Adds CSS styling to the homepage.

+++ db/migrate/20210603111854_create_destinations.rb
@@ -2,8 +2,8 @@ class CreateDestinations < ActiveRecord::Migration[6.1]
   def change
     create_table :destinations do |t|
       t.string :name
-      t.integer :distance
-      t.integer :gravity
+      t.float :distance
+      t.float :gravity
       t.integer :temperature
       t.string :image_url
       t.string :object_type
